Context: moved stats_info
